move stuff into Wallet component
